requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
uvloop>=0.19.0; platform_system != "Windows"
